requirements.txt
# Core dependencies
streamlit>=1.28.0
yt-dlp>=2023.10.13
webdav3>=3.14.6

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# AI/ML
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
uuid>=1.30
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
flake8>=6.1.0
pre-commit>=3.4.0
